Refactor `hasRunningMigration`
  - moved `hasRunningMigration` to **helpers.ts** so it is not tied
    to a component directly
  - **providerActions.tsx** changed to use the new import location
