+++ pkg/web/src/app/Providers/components/ProvidersTable/index.ts
@@ -1,1 +1,2 @@
 export { ProvidersTable } from './ProvidersTable';
+export { hasRunningMigration } from './helpers';
